移除未使用的导入和代码，优化项目结构
- 删除了多个类中的未使用导入语句，如Result、ResultCodeEnum等- 移除了部分接口和实现类中的多余代码
- 优化了部分方法参数类型，如从Long改为Integer
- 删除了UserServiceImpl中的多余注释

+++ web/src/main/java/com/mashiro/controller/RoleController.java
@@ -34,8 +34,10 @@ public class RoleController {
 
     @Operation(summary = "删除角色")
     @DeleteMapping("removeById")
-    public Result removeRole(@RequestParam Long id) {
+    public Result removeRole(@RequestParam Integer id) {
         roleService.removeById(id);
         return Result.ok();
     }
+
+    //获取角色所拥有的菜单路径
 }
